Using the fish method for sourcing
Apparently when you finally figure out how to make the default shell
fish you can't use source as that is a bash construct. Fish allows
you to use source if it isn't the default shell but doesn't allow
if it is the default shell. .... I have strong non positive feels
but it is fixed now so I am gettin back to work.